1) normalized crime, 2) changed crime fields to float, 3) modified score to have negatives, 4) Changed formatting, 5) added scale to choropleth

+++ foodmap/models.py
@@ -52,8 +52,8 @@ class EvaluationPoint(models.Model):
     population = models.FloatField(default=0)
     income = models.PositiveIntegerField(default=0)
 
-    crime_count_local = models.PositiveIntegerField(default=0)
-    crime_count_neighborhood = models.PositiveIntegerField(default=0)
+    crime_count_local = models.FloatField(default=0)
+    crime_count_neighborhood = models.FloatField(default=0)
 
     competitor_restaurant_count = models.PositiveSmallIntegerField(default=0)
     partner_restaurant_count = models.PositiveSmallIntegerField(default=0)
